Revert Next to `15.5.2`
