+++ package.json
@@ -98,7 +98,8 @@
     "@jest/types": "30.0.5",
     "form-data": "4.0.4",
     "on-headers": "1.1.0",
-    "react-is": "19"
+    "react-is": "19",
+    "sharp": "0.34.4"
   },
   "dependencies": {
     "@bbc/reverb-url-helper": "https://mybbc-analytics.files.bbci.co.uk/reverb-url-helper/bbc-reverb-url-helper-1.2.0.tgz",
