+++ ws-nextjs-app/package.json
@@ -5,15 +5,15 @@
   "scripts": {
     "setupDevEnv": "cp ../envConfig/local.env .env && sed 's/:7080/:7081/g' .env > .env.tmp && mv .env.tmp .env && mkdir -p public && cp -r ../public/. public",
     "moveStaticAssets": "mkdir -p build/standalone/ws-nextjs-app/public/_next/static && cp -r build/static/. build/standalone/ws-nextjs-app/public/_next/static && cp -r public/. build/standalone/ws-nextjs-app/public",
-    "build:local": "yarn setupDevEnv && NODE_ENV=production && next build --webpack && yarn moveStaticAssets",
-    "build:preview": "cp ../envConfig/preview1.env .env && NODE_ENV=production && next build --webpack",
-    "build:preview1": "cp ../envConfig/preview1.env .env && NODE_ENV=production && next build --webpack",
-    "build:preview2": "cp ../envConfig/preview2.env .env && NODE_ENV=production && next build --webpack",
-    "build:test": "cp ../envConfig/test.env .env && NODE_ENV=production && next build --webpack",
-    "build:live": "cp ../envConfig/live.env .env && NODE_ENV=production && next build --webpack",
+    "build:local": "yarn setupDevEnv && NODE_ENV=production && next build && yarn moveStaticAssets",
+    "build:preview": "cp ../envConfig/preview1.env .env && NODE_ENV=production && next build",
+    "build:preview1": "cp ../envConfig/preview1.env .env && NODE_ENV=production && next build",
+    "build:preview2": "cp ../envConfig/preview2.env .env && NODE_ENV=production && next build",
+    "build:test": "cp ../envConfig/test.env .env && NODE_ENV=production && next build",
+    "build:live": "cp ../envConfig/live.env .env && NODE_ENV=production && next build",
     "build": "yarn build:local",
-    "dev": "yarn setupDevEnv && next dev -p 7081 --webpack",
-    "dev-https": "yarn setupDevEnv && next dev -p 7081 --webpack --experimental-https",
+    "dev": "yarn setupDevEnv && next dev -p 7081",
+    "dev-https": "yarn setupDevEnv && next dev -p 7081 --experimental-https",
     "start": "NODE_ENV=production HOSTNAME=127.0.0.1 PORT=7081 node build/standalone/ws-nextjs-app/server.js",
     "stop": "lsof -t -i:7081 | xargs kill",
     "test": "NODE_OPTIONS=--no-experimental-strip-types jest --ci --colors --selectProjects='Unit Tests'",
@@ -36,12 +36,12 @@
     "@emotion/jest": "^11.14.2",
     "@jest/environment": "30.2.0",
     "@jest/types": "30.2.0",
-    "@next/swc-darwin-arm64": "16.0.1",
-    "@next/swc-darwin-x64": "16.0.1",
-    "@next/swc-linux-arm64-gnu": "16.0.1",
-    "@next/swc-linux-x64-gnu": "16.0.1",
-    "@next/swc-win32-arm64-msvc": "16.0.1",
-    "@next/swc-win32-x64-msvc": "16.0.1",
+    "@next/swc-darwin-arm64": "15.5.2",
+    "@next/swc-darwin-x64": "15.5.2",
+    "@next/swc-linux-arm64-gnu": "15.5.2",
+    "@next/swc-linux-x64-gnu": "15.5.2",
+    "@next/swc-win32-arm64-msvc": "15.5.2",
+    "@next/swc-win32-x64-msvc": "15.5.2",
     "@testing-library/jest-dom": "6.8.0",
     "@testing-library/react": "16.3.0",
     "babel-plugin-transform-rename-import": "2.3.0",
@@ -57,7 +57,7 @@
     "ts-node": "^10.9.2"
   },
   "dependencies": {
-    "next": "16.0.1",
+    "next": "15.5.2",
     "sharp": "0.34.4"
   }
 }
